December 15.
It took me a while to figure out the approach to part 2: That the uncovered spot must be just outside the edge of a sensor's "reach", i.e., the distance to its nearest sensor. This greatly reduces the number of potential spots to check.

Then it took me a bit of time to figure out that my first result to part 2 overflowed.